Bumps: Normalize type-casting in `bbp_bump_user_reply_count()` missed in r5732
See #2801


git-svn-id: https://bbpress.svn.wordpress.org/trunk@5888 9866e705-20ec-0310-96e7-cbb4277adcfb

+++ src/includes/users/functions.php
@@ -1839,7 +1839,7 @@ function bbp_bump_user_reply_count( $user_id = 0, $difference = 1 ) {
 	$user_reply_count = (int) ( $count + $difference );
 
 	// Add them up and filter them
-	$new_count = apply_filters( 'bbp_bump_user_reply_count', $user_reply_count, $user_id, $difference, $count );
+	$new_count = (int) apply_filters( 'bbp_bump_user_reply_count', $user_reply_count, $user_id, $difference, $count );
 
 	return bbp_update_user_reply_count( $user_id, $new_count );
 }
